+++ example/integration_test/plugin_integration_test.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:simple_local_auth/models/auth_results.dart';
 import 'package:simple_local_auth/simple_local_auth.dart';
 
 void main() {